+++ src/Movim/Console/ConfigCommand.php
@@ -115,6 +115,6 @@ class ConfigCommand extends Command
             }
         }
 
-        return 0;
+        return Command::SUCCESS;
     }
 }